feat: migrate String monkey patch to BT specific module and include in String
