--- lib/string/emoji.rb
@@ -1,10 +1,0 @@
-class String
-  def strip_emojis
-    gsub(Unicode::Emoji::REGEX, "")
-  end
-
-  def only_emoji?
-    return false if strip.empty?
-    strip_emojis.strip.empty?
-  end
-end
